openstack/requirements.txt
keystoneauth1
python-heatclient
